pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.74.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
    "selenium>=4.31.0",
]
